+++ index.js
@@ -7,9 +7,10 @@
 var zlib = require('zlib'),
     crypto = require('crypto'),
     through2 = require('through2'),
+    from2 = require('from2'),
     duplexer2 = require('duplexer2');
 
-var validHashOutputFormats = { 'hex': true, 'base64': true, 'buffer': true };
+var validHashOutputFormats = { 'hex': true, 'buffer': true };
 var validBlobTypes = { 'blob': true, 'tree': true, 'commit': true, 'tag': true };
 
 var blobWriter = function (options) {
@@ -108,4 +109,92 @@ var blobReader = function (options) {
   return duplexer2(inflator, inflator.pipe(transform));
 };
 
-module.exports = exports = { blobWriter: blobWriter, blobReader: blobReader };
+// Lifted from: https://github.com/creationix/js-git/blob/master/lib/modes.js
+// MIT Licensed
+var gitModes = {
+  isBlob: function (mode) {
+    return (mode & 0140000) === 0100000;
+  },
+  isFile: function (mode) {
+    return (mode & 0160000) === 0100000;
+  },
+  toType: function (mode) {
+    if (mode === 0160000) return "commit";
+    if (mode ===  040000) return "tree";
+    if ((mode & 0140000) === 0100000) return "blob";
+    return "unknown";
+  },
+  tree:    040000,
+  blob:   0100644,
+  file:   0100644,
+  exec:   0100755,
+  sym:    0120000,
+  commit: 0160000
+};
+
+// Lifted from: https://github.com/creationix/js-git/blob/master/lib/object-codec.js
+// MIT Licensed
+function treeMap(key) {
+  var entry = this[key];
+  return {
+    name: key,
+    mode: entry.mode,
+    hash: entry.hash
+  };
+}
+
+// Lifted from: https://github.com/creationix/js-git/blob/master/lib/object-codec.js
+// MIT Licensed
+function treeSort(a, b) {
+  var aa = (a.mode === gitModes.tree) ? a.name + "/" : a.name;
+  var bb = (b.mode === gitModes.tree) ? b.name + "/" : b.name;
+  return aa > bb ? 1 : aa < bb ? -1 : 0;
+}
+
+// Adapted from: https://github.com/creationix/js-git/blob/master/lib/object-codec.js
+// MIT Licensed
+var treeWriter = function (body, options) {
+  if (options) {
+    if (typeof options !== 'object') {
+      console.error('Bad options object passed to blobReader');
+      return null;
+    }
+  } else {
+    options = {};
+  }
+  var treeBuffers = [];
+  var tbSize = 0;
+  if (Array.isArray(body)) throw new TypeError("Tree must be in object form");
+  var list = Object.keys(body).map(treeMap, body).sort(treeSort);
+  for (var i = 0, l = list.length; i < l; i++) {
+    var entry = list[i];
+    if (typeof entry.hash === 'string') {
+      entry.hash = new Buffer(entry.hash, 'hex');
+    }
+    treeBuffers[2*i] = new Buffer(entry.mode.toString(8) + " " + entry.name + "\0");
+    treeBuffers[2*i+1] = entry.hash;
+    tbSize += treeBuffers[2*i].length + treeBuffers[2*i+1].length;
+  }
+  var buff = Buffer.concat(treeBuffers, tbSize);
+  var last = 0;
+  options.size = tbSize;
+  options.type = 'tree';
+  read = function (size, next) {
+    if (last >= tbSize) {
+      process.nextTick(function () { next(null, null); });
+    } else {
+      var stop = last+size;
+      var sl = buff.slice(last,stop);
+      last = stop;
+      process.nextTick(function () { next(null, sl); });
+    }
+  };
+  return from2(read).pipe(blobWriter(options));
+};
+
+module.exports = exports = {
+  blobWriter: blobWriter,
+  blobReader: blobReader,
+  treeWriter: treeWriter,
+  gitModes: gitModes
+};